+++ Controls/_popup/Page/Controller.ts
@@ -16,6 +16,7 @@ interface IPagePopupOptions extends IBasePopupOptions {
 }
 
 class PageController {
+    private _pageTemplate: string;
 
     /**
      * Получение опций окна для открытия страницы
@@ -24,10 +25,13 @@ class PageController {
      */
     getPagePopupOptions(pageId: string, popupOptions: IBasePopupOptions): Promise<unknown> {
         const resultPopupOptions = {...popupOptions};
+        if (!this._pageTemplate) {
+            throw new Error('На приложении не задан шаблон отображения страницы в окне');
+        }
         return DSPageController.getPageConfig(pageId).then((pageData) => {
             const templateOptions = this._getTemplateOptions(pageData, resultPopupOptions);
             templateOptions.prefetchResult = DSPageController.loadData(pageData, templateOptions.pageTemplateOptions);
-            resultPopupOptions.template = 'Controls/popupTemplate:Page';
+            resultPopupOptions.template = this._pageTemplate;
             resultPopupOptions.templateOptions = templateOptions;
             return resultPopupOptions;
         }).catch(() => {
@@ -41,6 +45,14 @@ class PageController {
      */
     loadModules(template: string): Promise<Control> {
         return loadModule(template);
+    }
+
+    /**
+     * Устанавливает шаблон отображения страницы на попапе
+     * @param template
+     */
+    setPageTemplate(template: string): void {
+        this._pageTemplate = template;
     }
 
     /**
