+++ Controls/dataSource.ts
@@ -37,7 +37,8 @@ export {default as NavigationController} from './_dataSource/NavigationControlle
 export {
     default as DataLoader,
     ILoadDataCustomConfig,
-    TLoadersConfigsMap as TLoadConfig
+    TLoadersConfigsMap as TLoadConfig,
+    TLoadResultMap
 } from './_dataSource/DataLoader';
 export {default as PageController, IPageConfig} from './_dataSource/PageController';
 export {getState as getControllerState, saveState as saveControllerState} from 'Controls/_dataSource/Controller/State';
